backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
drf-spectacular>=0.27.2
PyMySQL>=1.1.0
mysqlclient>=2.2.0
pillow>=10.0.0
pydantic>=2.4.2
langchain>=0.0.350
langchain-dashscope>=0.0.5
# pip install -r requirements.txt
# python manage.py runserver
# npm run dev